Crate libkavasam

Modules

errors

id

Structs

ReportMessage

A message sent to report an identifer

SignedHash

Represents a signed hash message, the hash being the digest of the identifier(phone number,email address, etc.)

SignedHashAsciiArmored

ASCII armored representation of SignedHash

Enums

Code

Supported hasing algorithms

IDType

Identier type supported by Kavasam

Type Definitions

Multihash

A Multihash with the same allocated size as the Multihashes produces by this derive.